+++ server/app.ts
@@ -122,7 +122,7 @@ export async function getApp() {
         req.path !== '/' &&
         req.path !== '/inscription' &&
         req.path !== '/connexion' &&
-        req.path !== '/professeur' &&
+        req.path !== '/login' &&
         req.path !== '/user-verified' &&
         req.path !== '/reset-password' &&
         req.path !== '/update-password'
@@ -135,7 +135,7 @@ export async function getApp() {
         req.user &&
         (req.path === '/inscription' ||
           req.path === '/connexion' ||
-          req.path === '/professeur' ||
+          req.path === '/login' ||
           req.path === '/reset-password' ||
           req.path === '/update-password')
       ) {
