+++ src/pages/login.tsx
@@ -1,64 +1,30 @@
-import Image from 'next/image';
 import { useRouter } from 'next/router';
 import qs from 'query-string';
 import React from 'react';
 
-import { NoSsr } from '@mui/material';
-import Backdrop from '@mui/material/Backdrop';
-import Button from '@mui/material/Button';
-import Checkbox from '@mui/material/Checkbox';
-import CircularProgress from '@mui/material/CircularProgress';
-import FormControlLabel from '@mui/material/FormControlLabel';
-import TextField from '@mui/material/TextField';
+import { Backdrop, Button, CircularProgress, Link, Tooltip } from '@mui/material';
 
-import { KeepRatio } from 'src/components/KeepRatio';
+import { KeepRatio } from '../components/KeepRatio';
+import { isRedirectValid } from '../components/accueil/NewHome';
 import { UserContext } from 'src/contexts/userContext';
-import Logo from 'src/svg/logo.svg';
-import { SSO_HOST, SSO_HOSTNAME, CLIENT_ID, onLoginSSO } from 'src/utils/sso';
-
-type User = {
-  username: string;
-  password: string;
-  remember: boolean;
-};
+import ArrowBack from 'src/svg/arrow_back.svg';
+import Logo from 'src/svg/logo_1village_classe.svg';
+import { onLoginSSO, SSO_HOST, CLIENT_ID } from 'src/utils/sso';
 
 const errorMessages = {
   0: 'Une erreur inconnue est survenue. Veuillez réessayer plus tard...',
-  1: 'Adresse e-mail ou pseudo invalide.',
-  2: 'Mot de passe invalide.',
-  3: 'Compte bloqué, trop de tentatives de connexion. Veuillez réinitialiser votre mot de passe.',
-  5: `Veuillez utiliser le login avec ${SSO_HOSTNAME} pour votre compte`,
-  6: 'Veuillez utiliser le login par email/mot de passe pour votre compte',
+  1: 'Identifiant invalides',
+  2: 'Compte bloqué, trop de tentatives de connexion. Veuillez réinitialiser votre mot de passe.',
 };
 
-const isRedirectValid = (redirect: string): boolean => {
-  // inner redirection.
-  if (redirect.startsWith('/')) {
-    return true;
-  }
-  // external, allow only same domain.
-  try {
-    const url = new URL(redirect);
-    return url.hostname.slice(-15) === '.parlemonde.org';
-  } catch {
-    return false;
-  }
-};
-
-const Login = () => {
+const SignInTeacher = () => {
   const router = useRouter();
-  const { login, loginWithSso } = React.useContext(UserContext);
+  const { loginWithSso } = React.useContext(UserContext);
   const redirect = React.useRef<string>('/');
   const [isLoading, setIsLoading] = React.useState(false);
-
-  const [user, setUser] = React.useState<User>({
-    username: '',
-    password: '',
-    remember: false,
-  });
   const [errorCode, setErrorCode] = React.useState(-1);
-
   const firstCall = React.useRef(false);
+
   const loginSSO = React.useCallback(
     async (code: string) => {
       if (firstCall.current === false) {
@@ -78,6 +44,7 @@ const Login = () => {
 
   React.useEffect(() => {
     const urlQueryParams = qs.parse(window.location.search);
+    console.log({ urlQueryParams });
     try {
       redirect.current = decodeURI((urlQueryParams.redirect as string) || '/');
     } catch (e) {
@@ -91,129 +58,105 @@ const Login = () => {
         setErrorCode(0);
       }
     }
-  }, [loginSSO]);
-
-  const updateUsername = (e: React.ChangeEvent<HTMLInputElement>): void => {
-    setUser((u) => ({ ...u, username: e.target.value }));
-    if (errorCode === 1) {
-      setErrorCode(-1);
-    }
-  };
-  const updatePassword = (e: React.ChangeEvent<HTMLInputElement>): void => {
-    setUser((u) => ({ ...u, password: e.target.value }));
-    if (errorCode === 1) {
-      setErrorCode(-1);
-    }
-  };
-
-  const onSubmit = async (e: React.FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    setIsLoading(true);
-    const response = await login(user.username, user.password, user.remember);
-    if (response.success) {
-      router.push(isRedirectValid(redirect.current) ? redirect.current : '/');
-    } else {
-      setErrorCode(response.errorCode || 0);
-    }
-    setIsLoading(false);
-  };
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
 
   return (
-    <div className="bg-gradiant">
-      <KeepRatio ratio={0.4} width="95%" maxWidth="1200px" minHeight="550px" className="login__container">
-        <div className="login__panel">
-          <form onSubmit={onSubmit} className="login__form">
-            <div className="flex-center" style={{ marginBottom: '0.8rem' }}>
-              <Logo style={{ width: '2.6rem', height: 'auto' }} />
-              <h1 className="title" style={{ margin: '0 0 0 0.5rem' }}>
-                1Village
-              </h1>
-            </div>
-            <p style={{ marginBottom: '1.5rem' }}>Se connecter</p>
-            {(errorCode === 0 || errorCode >= 3) && (
-              <p className="text text--small text--error text-center" style={{ margin: '0 0 1rem 0' }}>
-                {errorMessages[errorCode as 0] || errorMessages[0]}
-              </p>
-            )}
-            <NoSsr>
-              {SSO_HOST.length && CLIENT_ID && (
-                <>
-                  <div className="text-center" style={{ marginBottom: '1rem' }}>
-                    <Button color="primary" variant="contained" size="small" onClick={onLoginSSO}>
-                      Se connecter avec {SSO_HOSTNAME}
-                    </Button>
-                  </div>
-                  <div className="login__divider">
-                    <div className="login__or">
-                      <span style={{ fontSize: '1.2rem', padding: '0.25rem', backgroundColor: 'white' }}>OU</span>
-                    </div>
-                  </div>
-                </>
-              )}
-            </NoSsr>
-            <TextField
-              variant="standard"
-              label="Adresse email"
-              placeholder="Entrez votre adresse email"
-              name="username"
-              value={user.username}
-              fullWidth
-              onChange={updateUsername}
-              style={{ marginBottom: '1.5rem' }}
-              error={errorCode === 1}
-              helperText={errorCode === 1 ? errorMessages[1] : null}
-              InputLabelProps={{ shrink: true }}
-            />
-            <TextField
-              variant="standard"
-              label="Mot de passe"
-              placeholder="Entrez votre mot de passe"
-              name="password"
-              type="password"
-              value={user.password}
-              fullWidth
-              onChange={updatePassword}
-              style={{ marginBottom: '0.5rem' }}
-              error={errorCode === 2}
-              helperText={errorCode === 2 ? errorMessages[2] : null}
-              InputLabelProps={{ shrink: true }}
-            />
-            <div className="full-width">
-              <FormControlLabel
-                label="Se souvenir de moi"
-                style={{ marginBottom: '0.5rem', cursor: 'pointer' }}
-                control={
-                  <Checkbox
-                    color="primary"
-                    checked={user.remember}
-                    onChange={() => {
-                      setUser((u) => ({ ...u, remember: !u.remember }));
-                    }}
-                  />
-                }
-              />
-            </div>
-            <div className="text-center" style={{ marginBottom: '0.6rem' }}>
-              <Button type="submit" color="primary" variant="outlined">
+    <>
+      <div className="bg-gradiant" style={{ display: 'flex', flexDirection: 'column' }}>
+        <div
+          style={{
+            display: 'grid',
+            gridTemplateColumns: 'repeat(3, 1fr)',
+            gap: '10px',
+            background: 'white',
+            width: '95%',
+            maxWidth: '1200px',
+            borderRadius: '10px',
+            marginBottom: '2rem',
+          }}
+        >
+          {' '}
+          <Link
+            component="button"
+            variant="h3"
+            onClick={() => {
+              router.push('/');
+            }}
+            sx={{
+              placeSelf: 'flex-start',
+              marginRight: '1rem',
+              fontSize: '0.875rem',
+            }}
+          >
+            <Logo style={{ width: '10.563rem', height: 'auto', margin: '10px 0 5px 10px' }} />
+          </Link>
+          <h1 style={{ placeSelf: 'center' }}>Professeurs des écoles</h1>
+          <Link
+            component="button"
+            variant="h3"
+            onClick={() => {
+              router.push('/');
+            }}
+            sx={{
+              placeSelf: 'center end',
+              marginRight: '1rem',
+              fontSize: '0.875rem',
+            }}
+          >
+            <ArrowBack /> Village en famille
+          </Link>
+        </div>
+        <KeepRatio ratio={0.45} width="95%" maxWidth="1200px" minHeight="400px" className="login__container">
+          <div className="text-center" style={{ display: 'flex', flexDirection: 'column', alignItems: 'center', marginTop: '2rem' }}>
+            <iframe
+              src="https://player.vimeo.com/video/754287113?h=181d44f047"
+              width="640"
+              height="360"
+              frameBorder="0"
+              allow="autoplay; fullscreen"
+              allowFullScreen
+              style={{ height: '65%', width: '80%' }}
+            ></iframe>
+            {SSO_HOST.length && CLIENT_ID ? (
+              <Button color="primary" variant="outlined" style={{ marginTop: '0.8rem' }} onClick={onLoginSSO}>
                 Se connecter
               </Button>
-            </div>
-            <div className="text-center">
-              <a className="text text--small text--primary">Mot de passe oublié ?</a>
-            </div>
-          </form>
-        </div>
-        <div className="login__panel login__panel--with-blue-background" style={{ padding: '1rem' }}>
-          <div style={{ width: '100%', height: '100%', position: 'relative' }}>
-            <Image src="/family_values_best_friends.png" layout="fill" objectFit="contain" alt="image de login" />
+            ) : (
+              <Tooltip title="Vous devez posséder un compte enseignant pour vous connecter" arrow>
+                <span>
+                  <Button color="primary" variant="outlined" style={{ marginTop: '0.8rem' }} disabled>
+                    Se connecter
+                  </Button>
+                </span>
+              </Tooltip>
+            )}
+            <Link
+              href="https://prof.parlemonde.org/1village/"
+              rel="noreferrer"
+              target="_blank"
+              underline="none"
+              sx={{
+                fontSize: '0.875rem',
+                mt: 2,
+                fontWeight: 'bold',
+              }}
+            >
+              S&apos;inscrire
+            </Link>
+
+            {/* Bloc error text */}
+            <small style={{ color: 'tomato', fontWeight: 'bold' }}>
+              {errorCode !== -1 ? errorMessages[errorCode as 0] || errorMessages[0] : null}
+            </small>
           </div>
-        </div>
-      </KeepRatio>
-      <Backdrop style={{ zIndex: 2000, color: 'white' }} open={isLoading}>
-        <CircularProgress color="inherit" />
-      </Backdrop>
-    </div>
+        </KeepRatio>
+        <Backdrop style={{ zIndex: 2000, color: 'white' }} open={isLoading}>
+          <CircularProgress color="inherit" />
+        </Backdrop>
+      </div>
+    </>
   );
 };
 
-export default Login;
+export default SignInTeacher;
